requirements.txt
pyinstaller==4.0
twitchkeywords==1.0
asyncio==3.4.3
python-dotenv==0.13.0
twitchio==1.1.0
